Refactor preview API
