+++ x-pack/platform/plugins/shared/alerting/server/routes/rule/apis/execute/execute_rule_route.ts
@@ -6,19 +6,11 @@
  */
 
 import { schema } from '@kbn/config-schema';
-import { v4 as uuidv4 } from 'uuid';
 import type { RouteOptions } from '../../..';
 import { createBodySchemaV1 } from '../../../../../common/routes/rule/apis/create';
-import type {
-  RuleExecutorOptions,
-  RuleExecutorServices,
-  AsyncSearchParams,
-  AsyncSearchStrategies,
-} from '../../../../types';
-import { BASE_ALERTING_API_PATH, DISABLE_FLAPPING_SETTINGS } from '../../../../types';
+import { BASE_ALERTING_API_PATH } from '../../../../types';
 import { DEFAULT_ALERTING_ROUTE_SECURITY } from '../../../constants';
 import { verifyAccessAndContext } from '../../../lib';
-import { createEsClientWithMeta } from './es_client_meta_wrapper';
 
 export const executeRuleRoute = ({ router, licenseState, core }: RouteOptions) => {
   router.post(
@@ -51,12 +43,19 @@ export const executeRuleRoute = ({ router, licenseState, core }: RouteOptions) =
     router.handleLegacyErrors(
       verifyAccessAndContext(licenseState, async function (context, req, res) {
         try {
+          const requestParams = req.body;
           const coreContext = await context.core;
-          const [, { data, dataViews, share }] = await core.getStartServices();
           const alertingContext = await context.alerting;
           const rulesClient = await alertingContext.getRulesClient();
-          const requestParams = req.body;
           const ruleType = alertingContext.getType(requestParams.rule_type_id);
+          const [, { data, dataViews, share }] = await core.getStartServices();
+          const savedObjectsClient = coreContext.savedObjects.client;
+          const dataViewsService = await dataViews.dataViewsServiceFactory(
+            savedObjectsClient,
+            coreContext.elasticsearch.client.asInternalUser
+          );
+          const searchSourceClient = await data.search.searchSource.asScoped(req);
+          const uiSettingsClient = coreContext.uiSettings.client;
 
           if (!ruleType) {
             return res.badRequest({
@@ -64,85 +63,112 @@ export const executeRuleRoute = ({ router, licenseState, core }: RouteOptions) =
             });
           }
 
-          const esQueries: any[] = [];
-          const asCurrentUserWithMeta = createEsClientWithMeta(
-            coreContext.elasticsearch.client.asCurrentUser,
-            esQueries
-          );
-
-          const savedObjectsClient = coreContext.savedObjects.client;
-          const dataViewsService = await dataViews.dataViewsServiceFactory(
+          const esQueries = await rulesClient.preview({
+            requestParams: req.body,
+            ruleType,
+            esClient: coreContext.elasticsearch.client.asCurrentUser,
             savedObjectsClient,
-            coreContext.elasticsearch.client.asInternalUser
-          );
-          const searchSourceClient = await data.search.searchSource.asScoped(req);
-
-          const services: RuleExecutorServices = {
-            alertsClient: {
-              isTrackedAlert: () => false,
-              getAlertLimitValue: () => 1000,
-              setAlertLimitReached: () => {},
-              getRecoveredAlerts: () => [],
-              report: () => {},
-            },
-            getDataViews: async () => dataViewsService,
-            getMaintenanceWindowIds: async () => [],
-            getSearchSourceClient: async () => searchSourceClient,
-            savedObjectsClient: coreContext.savedObjects.client,
-            scopedClusterClient: {
-              asCurrentUser: asCurrentUserWithMeta,
-            },
+            dataViewsService,
+            searchSourceClient,
+            getSpaceId: rulesClient.getSpaceId,
             share,
-            shouldStopExecution: () => false,
-            shouldWriteAlerts: () => false,
-            uiSettingsClient: coreContext.uiSettings.client,
-            getAsyncSearchClient: <T extends AsyncSearchParams>(
-              strategy: AsyncSearchStrategies
-            ) => {
-              throw new Error('Not implemented');
-            },
-          };
-
-          const rule = {
-            id: uuidv4(),
-            ...requestParams,
-            actions: requestParams.actions || [],
-            schedule: requestParams.schedule || { interval: '1m' },
-            enabled: requestParams.enabled === undefined ? true : requestParams.enabled,
-            name: requestParams.name || 'On-demand execution',
-            tags: requestParams.tags || [],
-            throttle: requestParams.throttle || null,
-            createdBy: 'elastic',
-            updatedBy: 'elastic',
-            createdAt: new Date(),
-            updatedAt: new Date(),
-            apiKeyOwner: 'elastic',
-            notifyWhen: 'onActionGroup',
-            consumer: requestParams.consumer,
-            rule_type_id: requestParams.rule_type_id,
-            params: requestParams.params,
-          };
-
-          const options: RuleExecutorOptions = {
-            executionId: uuidv4(),
-            logger: context.logger,
-            params: requestParams.params,
-            previousStartedAt: null,
-            rule,
-            services,
-            spaceId: rulesClient.getSpaceId(),
-            startedAt: new Date(),
-            startedAtOverridden: false,
-            state: {},
-            namespace: rulesClient.getSpaceId(),
-            flappingSettings: DISABLE_FLAPPING_SETTINGS,
-            getTimeRange: () => ({ from: 'now-5m', to: 'now' }),
-            isServerless: false,
-          };
-
-          await ruleType.executor(options);
+            uiSettingsClient,
+          });
 
           return res.ok({ body: { _inspect: { esQueries } } });
+
+          // const coreContext = await context.core;
+          // const [, { data, dataViews, share }] = await core.getStartServices();
+          // const alertingContext = await context.alerting;
+          // const rulesClient = await alertingContext.getRulesClient();
+          // const requestParams = req.body;
+          // const ruleType = alertingContext.getType(requestParams.rule_type_id);
+          //
+          // if (!ruleType) {
+          //   return res.badRequest({
+          //     body: `Rule type '${requestParams.rule_type_id}' is not registered.`,
+          //   });
+          // }
+          //
+          // const esQueries: any[] = [];
+          // const asCurrentUserWithMeta = createEsClientWithMeta(
+          //   coreContext.elasticsearch.client.asCurrentUser,
+          //   esQueries
+          // );
+          //
+          // const savedObjectsClient = coreContext.savedObjects.client;
+          // const dataViewsService = await dataViews.dataViewsServiceFactory(
+          //   savedObjectsClient,
+          //   coreContext.elasticsearch.client.asInternalUser
+          // );
+          // const searchSourceClient = await data.search.searchSource.asScoped(req);
+          //
+          // const services: RuleExecutorServices = {
+          //   alertsClient: {
+          //     isTrackedAlert: () => false,
+          //     getAlertLimitValue: () => 1000,
+          //     setAlertLimitReached: () => {},
+          //     getRecoveredAlerts: () => [],
+          //     report: () => {},
+          //   },
+          //   getDataViews: async () => dataViewsService,
+          //   getMaintenanceWindowIds: async () => [],
+          //   getSearchSourceClient: async () => searchSourceClient,
+          //   savedObjectsClient: coreContext.savedObjects.client,
+          //   scopedClusterClient: {
+          //     asCurrentUser: asCurrentUserWithMeta,
+          //   },
+          //   share,
+          //   shouldStopExecution: () => false,
+          //   shouldWriteAlerts: () => false,
+          //   uiSettingsClient: coreContext.uiSettings.client,
+          //   getAsyncSearchClient: <T extends AsyncSearchParams>(
+          //     strategy: AsyncSearchStrategies
+          //   ) => {
+          //     throw new Error('Not implemented');
+          //   },
+          // };
+          //
+          // const rule = {
+          //   id: uuidv4(),
+          //   ...requestParams,
+          //   actions: requestParams.actions || [],
+          //   schedule: requestParams.schedule || { interval: '1m' },
+          //   enabled: requestParams.enabled === undefined ? true : requestParams.enabled,
+          //   name: requestParams.name || 'On-demand execution',
+          //   tags: requestParams.tags || [],
+          //   throttle: requestParams.throttle || null,
+          //   createdBy: 'elastic',
+          //   updatedBy: 'elastic',
+          //   createdAt: new Date(),
+          //   updatedAt: new Date(),
+          //   apiKeyOwner: 'elastic',
+          //   notifyWhen: 'onActionGroup',
+          //   consumer: requestParams.consumer,
+          //   rule_type_id: requestParams.rule_type_id,
+          //   params: requestParams.params,
+          // };
+          //
+          // const options: RuleExecutorOptions = {
+          //   executionId: uuidv4(),
+          //   logger: context.logger,
+          //   params: requestParams.params,
+          //   previousStartedAt: null,
+          //   rule,
+          //   services,
+          //   spaceId: rulesClient.getSpaceId(),
+          //   startedAt: new Date(),
+          //   startedAtOverridden: false,
+          //   state: {},
+          //   namespace: rulesClient.getSpaceId(),
+          //   flappingSettings: DISABLE_FLAPPING_SETTINGS,
+          //   getTimeRange: () => ({ from: 'now-5m', to: 'now' }),
+          //   isServerless: false,
+          // };
+          //
+          // await ruleType.executor(options);
+          //
+          // return res.ok({ body: { _inspect: { esQueries } } });
         } catch (error) {
           console.error(error);
         }
